Fix issue (in most cases) of multiple monitors
Sometimes with multiple monitors and the same number of spaces
and the same space selected we won't be able to easily identify
which monitor you're on (and therefore which names we should use).
This adds in a fallback to figure out which monitor has the same
dimensions of your element and uses that.